refactor(view): move log panel rendering out of view.go
Log rendering carries its own scroll state, filter and highlight handling, and
line styling, all of which only matter to the log panels.

Declarations move unchanged.
